BaSP-MR-226: firebase backend configuration

+++ src/controllers/admins.js
@@ -1,68 +1,107 @@
 const Admin = require('../models/Admins');
+const firebaseApp = require('../helper/firebase');
 
-const updateAdmin = (req, res) => {
+const updateAdmin = async (req, res) => {
   const { id } = req.params;
   const {
-    firstName, lastName, dni, phone, email, city, password,
+    firstName, lastName, dni, phone, email, city,
   } = req.body;
 
-  Admin.findByIdAndUpdate(
-    id,
-    {
+  try {
+    const existingAdmin = await Admin.findOne({ _id: id });
+
+    if (!existingAdmin) {
+      return res.status(404).json({
+        message: 'This Admin does not exists',
+        data: null,
+        error: true,
+      });
+    }
+    const { firebaseUid } = existingAdmin;
+
+    await firebaseApp.auth().updateUser(firebaseUid, {
+      email: req.body.email,
+      password: req.body.password,
+    });
+
+    const result = await Admin.findByIdAndUpdate(id, {
       firstName,
       lastName,
       dni,
       phone,
       email,
       city,
-      password,
-    },
-    { new: true },
-  )
-    .then((result) => {
-      if (!result) {
-        return res.status(404).json({
-          message: `Admin with ID ${id} not found`,
-          data: null,
-          error: true,
-        });
-      }
-      return res.status(200).json({
-        message: 'Admin updated!',
-        data: result,
-        error: false,
+    }, { new: true });
+
+    if (!result) {
+      return res.status(404).json({
+        message: `The id ${id} was not found`,
+        data: null,
+        error: true,
       });
-    })
-    .catch((error) => res.status(500).json({
+    }
+    return res.status(200).json({
+      message: 'Admin Updated',
+      data: result,
+      error: false,
+    });
+  } catch (error) {
+    return res.status(500).json({
       message: error,
       data: null,
       error: true,
-    }));
+    });
+  }
 };
 
-const createAdmin = (req, res) => {
+const createAdmin = async (req, res) => {
   const {
-    firstName, lastName, dni, phone, email, city, password,
+    firstName, lastName, dni, phone, email, city,
   } = req.body;
-  Admin.create({
-    firstName,
-    lastName,
-    dni,
-    phone,
-    email,
-    city,
-    password,
-  })
-    .then((result) => res.status(201).json({
+
+  let firebaseUid;
+  try {
+    const existingAdmin = await Admin.findOne({ email });
+
+    if (existingAdmin) {
+      return res.status(400).json({
+        message: 'This email is already used',
+        data: null,
+        error: true,
+      });
+    }
+
+    const newFirebaseUser = await firebaseApp.auth().createUser({
+      email: req.body.email,
+      password: req.body.password,
+    });
+
+    firebaseUid = newFirebaseUser.uid;
+
+    await firebaseApp.auth().setCustomUserClaims(newFirebaseUser.uid, { role: 'ADMIN' });
+
+    const result = await Admin.create({
+      firebaseUid,
+      firstName,
+      lastName,
+      dni,
+      phone,
+      email,
+      city,
+    });
+
+    return res.status(201).json({
       message: 'Admin created',
       data: result,
       error: false,
-    }))
-    .catch((error) => res.status(500).json({
+    });
+  } catch (error) {
+    return res.status(500).json({
       message: error,
       data: null,
       error: true,
-    }));
+    });
+  }
 };
 const getAdmins = (req, res) => {
   Admin.find()
@@ -103,29 +142,42 @@ const getAdminsById = (req, res) => {
     }));
 };
 
-const deleteAdmin = (req, res) => {
+const deleteAdmin = async (req, res) => {
   const { id } = req.params;
+  try {
+    const existingAdmin = await Admin.findOne({ _id: id });
 
-  Admin.findByIdAndDelete(id)
-    .then((result) => {
-      if (!result) {
-        return res.status(404).json({
-          message: `Admin with ID ${id} not found`,
-          data: null,
-          error: true,
-        });
-      }
-      return res.status(200).json({
-        message: 'Admin deleted!',
+    if (!existingAdmin) {
+      return res.status(404).json({
+        message: 'This Admin does not exists',
         data: null,
-        error: false,
+        error: true,
       });
-    })
-    .catch((error) => res.status(500).json({
-      message: error,
+    }
+    const { firebaseUid } = existingAdmin;
+
+    await firebaseApp.auth().deleteUser(firebaseUid);
+
+    const result = await Admin.findByIdAndDelete(id);
+    if (!result) {
+      return res.status(404).json({
+        message: `Admin with ID ${id} not found`,
+        data: null,
+        error: true,
+      });
+    }
+    return res.status(200).json({
+      message: 'Admin deleted!',
+      data: null,
+      error: false,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      message: 'Server Error',
       data: null,
       error: true,
-    }));
+    });
+  }
 };
 
 module.exports = {
